# Optimize `Record()` to avoid extra allocations 