@howardjohn 2021-09-14
https://github.com/census-instrumentation/opencensus-go/pull/1270 fixes the build failure